Perform processing with double

+++ src/main/java/ch/cern/spark/metrics/analysis/Analysis.java
@@ -17,7 +17,7 @@ public abstract class Analysis extends Component {
         super(Type.ANAYLSIS, subClass, name);
     }
 
-    public abstract AnalysisResult process(Instant timestamp, Float value);
+    public abstract AnalysisResult process(Instant timestamp, double value);
 
 }
